refactor: 상위 utils 이름을 common으로 수정
