+++ batch/src/main/java/com/ImSnacks/NyeoreumnagiBatch/WeatherJobScheduler.java
@@ -1,6 +1,6 @@
 package com.ImSnacks.NyeoreumnagiBatch;
 
-import com.ImSnacks.NyeoreumnagiBatch.utils.params.WeatherJobParams;
+import com.ImSnacks.NyeoreumnagiBatch.common.params.WeatherJobParams;
 import lombok.RequiredArgsConstructor;
 import org.springframework.batch.core.Job;
 import org.springframework.batch.core.JobParameters;
